+++ Escon.SisctNET.Web/Controllers/CestController.cs
@@ -24,6 +24,11 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpGet]
         public IActionResult Index()
         {
+            if (!SessionManager.GetCestInSession().Equals(10))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 var login = SessionManager.GetLoginInSession();
@@ -49,6 +54,11 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpGet]
         public IActionResult Create()
         {
+            if (!SessionManager.GetCestInSession().Equals(10))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 return View();
@@ -63,6 +73,11 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpPost]
         public IActionResult Create(Model.Cest entity)
         {
+            if (!SessionManager.GetCestInSession().Equals(10))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 entity.Created = DateTime.Now;
@@ -80,6 +95,11 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpGet]
         public IActionResult Edit(int id)
         {
+            if (!SessionManager.GetCestInSession().Equals(10))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 var result = _service.FindById(id, GetLog(Model.OccorenceLog.Read));
@@ -94,6 +114,11 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpPost]
         public IActionResult Edit(int id, Model.Cest entity)
         {
+            if (!SessionManager.GetCestInSession().Equals(10))
+            {
+                return Unauthorized();
+            }
+
             try
             { 
                 var rst = _service.FindById(id, GetLog(Model.OccorenceLog.Read));
@@ -111,6 +136,11 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpGet]
         public IActionResult Delete(int id)
         {
+            if (!SessionManager.GetCestInSession().Equals(10))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 _service.Delete(id, GetLog(Model.OccorenceLog.Delete));
