+++ Escon.SisctNET.Web/Controllers/CstController.cs
@@ -25,6 +25,11 @@ namespace Escon.SisctNET.Web.Controllers
 
         public IActionResult Index()
         {
+            if (!SessionManager.GetCstInSession().Equals(23))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 var login = SessionManager.GetLoginInSession();
@@ -48,6 +53,11 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpGet]
         public IActionResult Create()
         {
+            if (!SessionManager.GetCstInSession().Equals(23))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 return View();
@@ -61,6 +71,11 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpPost]
         public IActionResult Create(Model.Cst entity)
         {
+            if (!SessionManager.GetCstInSession().Equals(23))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 entity.Created = DateTime.Now;
@@ -77,6 +92,11 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpGet]
         public IActionResult Edit(int id)
         {
+            if (!SessionManager.GetCstInSession().Equals(23))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 var result = _service.FindById(id, GetLog(Model.OccorenceLog.Read));
@@ -91,6 +111,11 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpPost]
         public IActionResult Edit(int id, Model.Cst entity)
         {
+            if (!SessionManager.GetCstInSession().Equals(23))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 entity.Updated = DateTime.Now;
@@ -107,6 +132,11 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpGet]
         public IActionResult Delete(int id)
         {
+            if (!SessionManager.GetCstInSession().Equals(23))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 _service.Delete(id, GetLog(Model.OccorenceLog.Delete));
@@ -121,6 +151,11 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpPost]
         public IActionResult UpdateStatus([FromBody] Model.UpdateActive updateActive)
         {
+            if (!SessionManager.GetCstInSession().Equals(23))
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 var entity = _service.FindById(updateActive.Id, GetLog(Model.OccorenceLog.Read));
