+++ Escon.SisctNET.Web/Controllers/ProfileController.cs
@@ -22,6 +22,10 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpGet]
         public IActionResult Index()
         {
+            if (!SessionManager.GetProfileInSession().Equals(1))
+            {
+                return Unauthorized();
+            }
             try
             {
                 var login = SessionManager.GetLoginInSession();
@@ -46,6 +50,10 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpGet]
         public IActionResult Create()
         {
+            if (!SessionManager.GetProfileInSession().Equals(1))
+            {
+                return Unauthorized();
+            }
             try
             {
                 return View();
@@ -59,6 +67,10 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpPost]
         public IActionResult Create(Model.Profile entity)
         {
+            if (!SessionManager.GetProfileInSession().Equals(1))
+            {
+                return Unauthorized();
+            }
             try
             {
                 entity.Created = DateTime.Now;
@@ -76,6 +88,10 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpGet]
         public IActionResult Edit(int id)
         {
+            if (!SessionManager.GetProfileInSession().Equals(1))
+            {
+                return Unauthorized();
+            }
             try
             {
                 var result = _service.FindById(id, GetLog(Model.OccorenceLog.Read));
@@ -90,6 +106,10 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpPost]
         public IActionResult Edit(int id, Model.Profile entity)
         {
+            if (!SessionManager.GetProfileInSession().Equals(1))
+            {
+                return Unauthorized();
+            }
             try
             {
                 var rst = _service.FindById(id, GetLog(Model.OccorenceLog.Read));
@@ -107,6 +127,10 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpGet]
         public IActionResult Delete(int id)
         {
+            if (!SessionManager.GetProfileInSession().Equals(1))
+            {
+                return Unauthorized();
+            }
             try
             {
                 _service.Delete(id, GetLog(Model.OccorenceLog.Delete));
@@ -121,6 +145,10 @@ namespace Escon.SisctNET.Web.Controllers
         [HttpPost]
         public IActionResult UpdateStatus([FromBody] Model.UpdateActive updateActive)
         {
+            if (!SessionManager.GetProfileInSession().Equals(1))
+            {
+                return Unauthorized();
+            }
             try
             {
                 var entity = _service.FindById(updateActive.Id, GetLog(Model.OccorenceLog.Read));
